fix: run on stable rust and update CI
The CI is now testing only the version that is specified in the `rust_toolchain` file.
Previously that version plus stable was tested, but now as we are on stable, there is
not need to run tests twice.

The `--all` flag for `cargo` is deprecated, instead one should use `--workspace`. This
commit also makes this adjustment.